Moving game to common
