+++ src/common/active-rooms.js
