+++ src/common/animation.js
