+++ src/common/goat-game.js
