+++ src/common/lib/goat-diagnostics.js
