+++ src/common/lib/goat-math.js
