+++ src/common/lib/goat-telemetry.js
